refactor and move to new file
